requirements.txt
google-api-python-client==2.101.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
SQLAlchemy==2.0.23
pytest==7.4.3